+++ orthanc_tools/__init__.py
@@ -4,10 +4,9 @@ from .orthanc_monitor import OrthancMonitor
 from .orthanc_test_db_populator import OrthancTestDbPopulator
 from .pacs_migrator import PacsMigrator
 from .orthanc_comparator import OrthancComparator
-from .scheduler import Scheduler
-from .old_files_deleter import OldFilesDeleter
 
 from .hl7Lib import *
+from .helpers import *
 
 # Set default logging handler to avoid "No handler found" warnings.
 import logging