+++ orthanc_tools/helpers/__init__.py
@@ -1,0 +1,2 @@
+from .scheduler import Scheduler
+from .old_files_deleter import OldFilesDeleter